refactor(runs): move artifact index summaries to common
