+++ src/commands/runs/common.rs
@@ -34,6 +34,32 @@ pub struct RunSummary {
     pub artifact_index: Option<RigRunArtifactIndex>,
 }
 
+pub fn run_summaries_with_artifact_indexes(
+    store: &ObservationStore,
+    run_records: Vec<RunRecord>,
+) -> homeboy::core::Result<Vec<RunSummary>> {
+    let rig_run_ids = run_records
+        .iter()
+        .filter_map(|run| (run.kind == "rig" && run.rig_id.is_some()).then(|| run.id.clone()))
+        .collect::<Vec<_>>();
+    let mut artifacts_by_run = store.list_artifacts_for_runs(&rig_run_ids)?;
+    Ok(run_records
+        .into_iter()
+        .map(|run| {
+            let artifacts = artifacts_by_run.remove(&run.id).unwrap_or_default();
+            run_summary_with_artifact_index(run, &artifacts)
+        })
+        .collect())
+}
+
+fn run_summary_with_artifact_index(run: RunRecord, artifacts: &[ArtifactRecord]) -> RunSummary {
+    let artifact_index = homeboy::core::rig::artifact_index_for_run_with_artifacts(&run, artifacts);
+    RunSummary {
+        artifact_index,
+        ..super::run_summary(run)
+    }
+}
+
 /// Convert a `--since <duration>` flag value into an RFC-3339 timestamp.
 ///
 /// Accepts the same forms as elsewhere in the runs surface (s, m, h, d).
